feat(auth): refactor authentication handling in Sidebar and add SignInButton component
