+++ src/components/Sidebar.tsx
@@ -1,4 +1,3 @@
-'use client'
 import Link from 'next/link'
 import {
   Sidebar,
@@ -13,12 +12,10 @@ import { FiMessageSquare } from "react-icons/fi";
 import { IoSettingsOutline } from "react-icons/io5";
 import { FaRegUser } from "react-icons/fa";
 import Image from "next/image";
-import { useAuth } from "react-oidc-context";
-import { Button } from "@/components/ui/button";
+import { SignInButton } from "@/components/auth/SignInButton";
 
 
 export function SidebarComponent() {
-  const auth = useAuth()
   const items = [
     { title: "Home", url: '/', icon: Home},
     { title: "Messages", url: '/messages', icon: FiMessageSquare},
@@ -26,23 +23,6 @@ export function SidebarComponent() {
     { title: "Profile", url:'/profile', icon: FaRegUser},
     { title: "Settings", url: '/register', icon: IoSettingsOutline},
   ]
-
-  const signout = async () => {
-    const confirm = window.confirm("Are you sure you want to sign out?");
-    if (!confirm) {
-      return
-    }
-    const clientId = process.env.NEXT_PUBLIC_COGNITO_CLIENT_ID;
-    const logoutUri = process.env.NEXT_PUBLIC_COGNITO_CALLBACK_URL;
-    const cognitoDomain = process.env.NEXT_PUBLIC_COGNITO_USER_POOL;
-    window.location.href = `${cognitoDomain}/logout?client_id=${clientId}&logout_uri=${encodeURIComponent(logoutUri!)}`;
-    await auth.removeUser()
-  }
-
-
-  if (auth.error) {
-    return <div>Encountering error... {auth.error.message}</div>;
-  }
 
   return (
     <>
@@ -75,8 +55,7 @@ export function SidebarComponent() {
                     </SidebarMenuItem>
                   );
                 })}
-                {!auth.user && <Button onClick={() => auth.signinRedirect()}>Log In</Button>}
-                {auth.user && <Button onClick={() => signout()}>Sign Out</Button>}
+              <SignInButton/>
               </SidebarMenu>
             </SidebarGroup>
           </SidebarContent>
